Refactor EGL backend
